This commit provides the basic functionality to download the FCC ham radio database, unzip it, combine the desired fields into a new file, and clean up.
The download is currently async, and the other steps will likely need
to be. I will likely split out the processing step into more steps when
the move to async is made.

The current format for the stored local database is very simple. It
may need to be updated for speed later (based on previous experience with
a Mac version of this program), but this will do for now.